% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_outputation.R
\name{multiout}
\alias{multiout}
\title{Perform multiple outputation}
\usage{
multiout(fn, M, data, id, leave.as.list = FALSE)
}
\arguments{
\item{fn}{a function to apply to each sub-data set}

\item{M}{the (positive integer) number of outputations to perform}

\item{data}{the data.frame}

\item{id}{a string of the name describing each cluster}

\item{leave.as.list}{a logical indicating whether the results should be left as a list}
}
\value{
A vector (or list of vectors in the case of leave.as.list = TRUE) containing the outputated estimates
}
\description{
Perform multiple outputation on a dataset to obtain estimates of generic parameters
}
